Day1/Exercise2/requirements.txt
flask==2.0.1
werkzeug==2.0.1
google-cloud-bigquery>=3.11.4
google-auth>=2.3.3
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.26.0
names==0.3.0
schedule==1.1.0